Update automate.sh
I use the output of $XDG_TEMPLATES_DIR to get the translated path and check if it already exists or not. If so, I'll just move the contents of the *Templates* folder; if not, I'll first create that folder and than do the former thing.
As you can see I deleted rsync as it isn't needed any longer.